refactor and break apart menu class
