+++ src/com/alienation/coregamefiles/gamefunctionclasses/SaveGame.java
@@ -1,0 +1,92 @@
+package com.alienation.coregamefiles.gamefunctionclasses;
+
+import com.alienation.coregamefiles.enums.Rooms;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.transform.OutputKeys;
+import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
+
+import java.io.File;
+import java.util.List;
+
+import static com.alienation.coregamefiles.charactersetc.Oxygen.getOxygen;
+import static com.alienation.coregamefiles.charactersetc.Player.*;
+import static com.alienation.coregamefiles.charactersetc.Player.getInventory;
+import static com.alienation.coregamefiles.hashmaps.AvailableItemsHashMap.getAvailableItemsMap;
+
+public class SaveGame {
+
+    /* -- Save the Game -- START */
+    // Save the Game
+    public static void saveGameDataToFile() {
+        DocumentBuilderFactory dbFactory = DocumentBuilderFactory.newInstance();
+        DocumentBuilder dBuilder;
+        try {
+            dBuilder = dbFactory.newDocumentBuilder();
+            Document doc = dBuilder.newDocument();
+            //add elements to Document
+            Element rootElement =
+                    doc.createElementNS("", "GameState");
+            //append root element to document
+            doc.appendChild(rootElement);
+
+            //append child elements to root element
+            rootElement.appendChild(getGameElements(doc,"CurrentHealth",String.valueOf(getHealth())));
+            rootElement.appendChild(getGameElements(doc,"CurrentOxygen",String.valueOf(getOxygen())));
+            rootElement.appendChild(getGameElements(doc,"CurrentWeapon",String.valueOf(getCurrentWeapon())));
+            rootElement.appendChild(getGameElements(doc,"CurrentRoom",String.valueOf(getCurrentRoom())));
+            rootElement.appendChild(getGameElements(doc,"TempRoom",String.valueOf(getTempRoom())));
+            rootElement.appendChild(getGameElements(doc,"PreviousRoom",String.valueOf(getPreviousRoom())));
+
+            //append inventory list to root element
+            rootElement.appendChild(getGameData(doc,"Inventory",getInventory()));
+
+            //append room available item list to root element
+            for (Rooms room : Rooms.values()) {
+                rootElement.appendChild(getGameData(doc,room.toString(),getAvailableItemsMap().get(room)));
+            }
+
+            //for output to file, console
+            TransformerFactory transformerFactory = TransformerFactory.newInstance();
+            Transformer transformer = transformerFactory.newTransformer();
+            //for pretty print
+            transformer.setOutputProperty(OutputKeys.INDENT, "yes");
+            DOMSource source = new DOMSource(doc);
+
+            //write to console or file
+            //StreamResult console = new StreamResult(System.out);
+            StreamResult file = new StreamResult(new File(System.getProperty("user.dir") + "\\SaveState.xml"));
+
+            //write data
+            //transformer.transform(source, console);
+            transformer.transform(source, file);
+            System.out.println("We saved the game status!!");
+            System.exit(0);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    private static Node getGameData(Document doc, String element, List<String> items) {
+        Element data = doc.createElement(element);
+        for(String value : items){
+            data.appendChild(getGameElements(doc, "Item", value));
+        }
+        return data;
+    }
+
+    //utility method to create text node
+    private static Node getGameElements(Document doc, String name, String value) {
+        Element node = doc.createElement(name);
+        node.appendChild(doc.createTextNode(value));
+        return node;
+    }
+    /* -- Save the Game -- END */
+}
